porting to clang 10: - making CMakeLists.txt useful - remove lots of warnings - std::exception doesn't const char * constructor - missing includes - TODO: files are randomly spready and counts with include path to all directories - don't use `return std::move(...)` it blocks RVO - a lots of unused variables - IFDEFed windows include - removed <range> include - concepts result is not a type but another concept, as there is missing <concepts>  in clang10, I implemented the is_same - reordering constructor paramters - std::memory_order_relaxed proper name (as the memory_order enum is not yet implemented in libc++) - I don't know why there are GetVoxelRaw({expression})

+++ src/color_palette.cpp
@@ -82,13 +82,13 @@ Color ColorPalette::Get(Colors colorName)
 Color ColorPalette::GetPrimary(TankColor index)
 {
     assert(index >= 0 && index < PrimaryColors);
-    return Primaries[index];
+    return Primaries[static_cast<int>(index)];
 }
 
 Color * ColorPalette::GetTank(TankColor index)
 {
     assert(index >= 0 && index < PrimaryColors);
-    return Tanks[index];
+    return Tanks[static_cast<int>(index)];
 }
 
 void ColorPalette::Set(Colors colorName, Color color) { Values[static_cast<int>(colorName)] = color; }
@@ -96,13 +96,13 @@ void ColorPalette::Set(Colors colorName, Color color) { Values[static_cast<int>(
 void ColorPalette::SetPrimary(TankColor index, Color color)
 {
     assert(index >= 0 && index < PrimaryColors);
-    Primaries[index] = color;
+    Primaries[static_cast<int>(index)] = color;
 }
 
 void ColorPalette::SetTank(TankColor index, Color color_1, Color color_2, Color color_3)
 {
     assert(index >= 0 && index < PrimaryColors);
-    Tanks[index][0] = color_1;
-    Tanks[index][1] = color_2;
-    Tanks[index][2] = color_3;
+    Tanks[static_cast<int>(index)][0] = color_1;
+    Tanks[static_cast<int>(index)][1] = color_2;
+    Tanks[static_cast<int>(index)][2] = color_3;
 }
